# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 4.0.0
  description: TIAMO BURGER — API del Sistema POS
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Iniciar sesión
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsuarioPublico"
        "401":
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Obtener usuario autenticado
      responses:
        "200":
          description: Usuario actual
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsuarioPublico"
        "401":
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /caja/abrir:
    post:
      operationId: abrirCaja
      tags: [caja]
      summary: Abrir sesión de caja
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AbrirCajaInput"
      responses:
        "201":
          description: Caja abierta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CajaSesion"
  /caja/cerrar:
    post:
      operationId: cerrarCaja
      tags: [caja]
      summary: Cerrar sesión de caja con conteo de efectivo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CerrarCajaInput"
      responses:
        "200":
          description: Caja cerrada con resumen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumenCierre"
...
  /upload:
    post:
      operationId: subirImagen
      tags: [productos]
      summary: Subir imagen de producto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                imageData:
                  type: string
              required: [imageData]
      responses:
        "200":
          description: URL de la imagen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"
...
  /promociones:
    get:
      operationId: getPromociones
      tags: [productos]
      summary: Listar promociones
      responses:
        "200":
          description: Lista de promociones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Promocion"
    post:
      operationId: crearPromocion
      tags: [productos]
      summary: Crear promoción
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromocionInput"
      responses:
        "201":
          description: Promoción creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Promocion"
  /promociones/{id}:
    put:
      operationId: actualizarPromocion
      tags: [productos]
      summary: Actualizar promoción
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromocionInput"
      responses:
        "200":
          description: Promoción actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Promocion"
    delete:
      operationId: eliminarPromocion
      tags: [productos]
      summary: Eliminar promoción
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Eliminada
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    RolUsuario:
      type: string
      enum: [admin, mesero, cocinero, caja]
...
    UsuarioPublico:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
        usuario:
          type: string
        rol:
          $ref: "#/components/schemas/RolUsuario"
        activo:
          type: boolean
        creadoEn:
          type: string
          format: date-time
        fotoUrl:
          type: string
          nullable: true
      required: [id, nombre, usuario, rol, activo, creadoEn]
    LoginInput:
      type: object
      properties:
        usuario:
          type: string
        contrasena:
          type: string
      required: [usuario, contrasena]
...
    CajaSesion:
      type: object
      properties:
        id:
          type: integer
        usuarioId:
          type: integer
        usuarioNombre:
          type: string
        aperturaEn:
          type: string
          format: date-time
        cierreEn:
          type: string
          format: date-time
        montoInicial:
          type: integer
        totalEfectivo:
          type: integer
        totalTransferencia:
          type: integer
        totalPropinas:
          type: integer
        diferencia:
          type: integer
        estado:
          type: string
          enum: [abierta, cerrada]
        notas:
          type: string
      required: [id, usuarioId, usuarioNombre, aperturaEn, montoInicial, totalEfectivo, totalTransferencia, totalPropinas, diferencia, estado]
    AbrirCajaInput:
      type: object
      properties:
        montoInicial:
          type: integer
        notas:
          type: string
      required: [montoInicial]
    CerrarCajaInput:
      type: object
      properties:
        efectivoContado:
          type: integer
        notas:
          type: string
      required: [efectivoContado]
    ResumenCierre:
      type: object
      properties:
        id:
          type: integer
        totalEfectivo:
          type: integer
        totalTransferencia:
          type: integer
        totalPropinas:
          type: integer
        diferencia:
          type: integer
        pedidosCobrados:
          type: integer
        efectivoEsperado:
          type: integer
        efectivoContado:
          type: integer
        estado:
          type: string
      required: [id, totalEfectivo, totalTransferencia, totalPropinas, diferencia, pedidosCobrados, efectivoEsperado, efectivoContado, estado]
...
    Promocion:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
        descripcion:
          type: string
        tipo:
          type: string
          enum: [descuento, combo, 2x1, happy_hour]
        descuento:
          type: integer
        tipoDescuento:
          type: string
          enum: [porcentaje, fijo]
        activo:
          type: boolean
        diasSemana:
          type: string
        horaInicio:
          type: string
        horaFin:
          type: string
        imagenUrl:
          type: string
        creadoEn:
          type: string
          format: date-time
      required: [id, nombre, tipo, tipoDescuento, activo, creadoEn]
    PromocionInput:
      type: object
      properties:
        nombre:
          type: string
        descripcion:
          type: string
        tipo:
          type: string
          enum: [descuento, combo, 2x1, happy_hour]
        descuento:
          type: integer
        tipoDescuento:
          type: string
          enum: [porcentaje, fijo]
        activo:
          type: boolean
        diasSemana:
          type: string
        horaInicio:
          type: string
        horaFin:
          type: string
        imagenUrl:
          type: string
      required: [nombre]
    UploadResult:
      type: object
      properties:
        url:
          type: string
      required: [url]
...