# lib/api-zod/src/generated/types/rolUsuario.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TIAMO BURGER — API del Sistema POS
 * OpenAPI spec version: 4.0.0
 */

export type RolUsuario = (typeof RolUsuario)[keyof typeof RolUsuario];

export const RolUsuario = {
  admin: "admin",
  mesero: "mesero",
  cocinero: "cocinero",
  caja: "caja",
} as const;